Refactor extract functions from 'includes' Operator to utils
